+++ CHANGELOG.md
@@ -16,7 +16,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 ### Fixed
 
 - Remove non-functional `globalId` argument definition from `@update` https://github.com/nuwave/lighthouse/pull/1660
-- Resolve field middleware directives in lexical order
+- Resolve field middleware directives in lexical order https://github.com/nuwave/lighthouse/pull/1666
 
 ## 5.0.2